Initial commit of old code I had already written
Previously written with intent to upload as part of the main RuneLite
repo, now for a new repo (this repo) to work with the RuneLite plugin hub.

This was copied across with as few changes to compile and work as
possible (the plugin to work at least, the tests fail). So some things to
remember to look at:

- Some tests don't pass (in BankMemoryPluginTest)
- TODO/planning comments in test files
- Probably TODOs elsewhere
- Put code in a style that I like (not the RuneLite repo style)
- I should learn how the build.gradle should be laid out/formed
- Probably some minor nicety refactorings
- Update dependencies? The version numbers used here may be partly just
  the newest version available and partly the versions that were in my old
  checkout of the RuneLite repo.