On plugin startup, rememer to update bank info on client thread
...instead of on the EDT, as I was accidentally doing. This stops a bug
which happens when enabling the plugin after being already logged in
(on a character with saved bank info). The plugin would crash in the
startUp method so just not work. This only happen if enabling when
already logged in, else it went down a different code path which didn't
trigger the threading error.