Replace needless #repaint calls with one #validate call
This actually gets the item list to display properly on the first time,
fixing the bug where (when the item list contents is first set) the item
list appears as empty and requires the user to switch to a different
plugin panel and then back to this one. I still don't know why I need to
manually call #validate for it to be layed out properly, but whatever.