Rewrite how data is stored to use JSON
Instead of using a custom parser to serialise/deserialise bank save
data, use Gson to de/serialise it to JSON instead. This'll be better
for edge case stuff I now don't have to deal with (e.g. worrying
about escaping characters in logins and user entered save names which
will be added soon).

Gson is certainly slower though, so I kept ItemDataParser to
de/serialise the list of items in the bank as that part is 90% of the
parse time and ItemDataParser takes about 20~% as long compared to just
letting Gson do it itself.