Obscure user login emails in logs and UI
If a display name for an account with an email login couldn't be found,
then the email would be displayed in the UI. So... stop doing that.
Instead, display a censored version of the email (first 3 characters, or
less, and email length).

There's no explanation given to the user for why they see these weird
censored strings but it shouldn't really ever happen anyway (only if the
user messed with the JSON in the config). Besides, it'll automatically
fix as they log into their accounts and the display names get populated
again.

Certain errors also printed the username to the logs. Those should now
be omitted or censored.