Fix bug where current bank panel could be accidentally left reset
Bug was introduced with commit:

	b73626ae19d7ddd1fa26b20e4cd84df2aa3a4bd6
	Differentiate bank saves for special mode worlds

The issue was with CurrentBankPanelController:56. The result of the
 #equalsIgnoreCase call should have been negated but was not. Because of
this, the items list could be reset even if there was nothing new to
display. And when there's nothing new to display, the panel was left
blank until the items list would be updated.

This changes bring the call to #reset closer to where the items lists is
updated and links the logic a bit more, hopefully making the code
resistant to such a bug in the future (at least here).

This is probably a sign that I should start writing more tests...